requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
contourpy==1.2.1
cycler==0.12.1
efficientnet-pytorch==0.6.3
fonttools==4.53.1
google-auth==1.31.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
importlib_resources==6.4.3
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.5.4
Markdown==3.3.4
matplotlib==3.9.2
munch==2.5.0
numpy==1.23.0
oauthlib==3.1.1
packaging==24.1
Pillow==8.2.0
pretrainedmodels==0.7.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.1.2
python-box==7.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.5.0
timm==0.3.2
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.20.0
